[#79] chore: Product Item 스타일 분리 및 rename
